+++ src/main/scala/edu/rit/cs/mmior/jsonoid/generator/NumberGenerator.scala
@@ -3,6 +3,7 @@ package edu.rit.cs.mmior.jsonoid.generator
 import edu.rit.cs.mmior.jsonoid.discovery.schemas.{
   MaxNumValueProperty,
   MinNumValueProperty,
+  NumMultipleOfProperty,
   NumberSchema
 }
 
@@ -10,6 +11,12 @@ import org.json4s._
 
 object NumberGenerator extends Generator[NumberSchema, JDecimal] {
   def generate(schema: NumberSchema): JDecimal = {
+    val multiple = schema.properties
+      .getOrNone[NumMultipleOfProperty]
+      .flatMap(_.multiple)
+      .map(_.toDouble)
+      .getOrElse(0.1)
+
     val minValue = schema.properties
       .get[MinNumValueProperty]
       .minNumValue
@@ -21,8 +28,12 @@ object NumberGenerator extends Generator[NumberSchema, JDecimal] {
       .map(_.doubleValue)
       .getOrElse(minValue + 1000.0)
 
-    val range = maxValue - minValue
+    // Get the possible range of multiples
+    val range = (maxValue / multiple).floor - (minValue / multiple).ceil
 
-    JDecimal(util.Random.nextFloat * range + minValue)
+    // Find the smallest value in range which meets the necessary multiple
+    val smallest = (minValue / multiple).ceil * multiple
+
+    JDecimal((util.Random.nextFloat * range).toInt * multiple + smallest)
   }
 }